refactor: store 파일로 분리
