+++ src/components/chat/Course.tsx
@@ -3,7 +3,7 @@ import type { Course, Place } from "../../types/domain";
 import { useEffect, useState } from "react";
 import Button from "../common/Button";
 import usePostSaveTravel from "../../hooks/usePostSaveTravel";
-import { useChatStore } from "../../stores";
+import { useChatStore } from "../../stores/useChatStore";
 import { useNavigate } from "react-router";
 import ReactMarkdown from "react-markdown";
 import rehypeRaw from "rehype-raw";
@@ -94,9 +94,12 @@ const Course = ({ scrollDown, messageId, courses }: CourseProps) => {
         }
       }, 10);
 
-      return () => clearInterval(interval);
+      return () => {
+        clearInterval(interval);
+      };
+    } else {
+      localStorage.setItem("lastMessageId", "course_done");
     }
-    scrollDown();
   }, [courseIndex, charIndex, courses, scrollDown]);
 
   return (
