+++ src/components/chat/Region.tsx
@@ -1,7 +1,8 @@
 import React from "react";
 import ChatGroup from "./ChatGroup";
 import { isRegionType } from "../../utils";
-import { useChatStore, useTravelStore } from "../../stores";
+import { useTravelStore } from "../../stores/useTravelStore";
+import { useChatStore } from "../../stores/useChatStore";
 import Map from "./Map";
 
 const Region = () => {
